requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
argh==0.28.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==4.1.7
django-celery-beat==2.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
flower==1.2.0
frozenlist==1.3.3
humanize==4.6.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pathtools==0.1.2
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
